[MOV] mail: move values tracking in a new mixin
Some models (try to) track value changes without really inheriting
or using mail.thread capabilities. This is for example the case with
move lines, but also hr card wizards, payroll parameter lines, ...

Tracking value changes and logging on another model is a frequent
need on technical models. Submodels like lines often store value
users may want to track at a higher level.

In order to enable value tracking and managing them without inheriting
from the whole mail.thread stack we add a new mixin dealing with the
base computation of values tracking. Mail Thread then adds automatic
change computation, chatter management, message posting, ... on top
of it.

To ease 'mail.thread' code addition, '_track_finalize' now calls
an inner '_track_execute' receiving computed trackings and changes.
This allows to implement override-specific behavior only without
worrying on actual changes computation.

Future commits will move some additional features to the mixin. This
is done in several steps to ease readability and control feature move.

This does not change anything for most models as mail.thread inherits
from the new one.

Task-5935695 ([mail, various] Cleanup and test tracking usage)
Prepares Task-3645865 ([mail] In-body tracking)
